8576: clean up components into their own files
